Make bootstrap-collapse change maxHeight instead of height
The existing code sets the destination height to the calcualted
scrollHeight of the collapsed region. As a result, any further content
added to the collapsed region after it has been opened is forced into the
hidden part of that region becuase the height is not dynamically updated.

Since it's not possible to transition to a height of 'auto' with native
transitions, we have to instead set the height to auto, and transition the
maxHeight value. We can then clear the maxHeight value after the transition
completes to allow the collapsed region to make use of the height: auto
setting to expand for new content.